Update on "[TensorExpr] Pull changes from bertmaher/pytorch_fusion."
This PR (hopefully the last one of such kind) is merging changes from a
side branch where tensor expessions based fuser work has been done so
far. This PR is is a squashed version of changes in the side branch,
which is available here: https://github.com/bertmaher/pytorch

Differential Revision: [D20478208](https://our.internmc.facebook.com/intern/diff/D20478208)

[ghstack-poisoned]

+++ torch/csrc/jit/tensorexpr/hash_server.h
@@ -1,4 +1,3 @@
-
 #include <torch/csrc/jit/tensorexpr/ir.h>
 #include <torch/csrc/jit/tensorexpr/ir_printer.h>
 #include <torch/csrc/jit/tensorexpr/ir_visitor.h>
@@ -123,7 +122,7 @@ class HashProvider : public IRVisitor {
         v,
         hash_combine(
             hashOf(v->lhs()),
-            v->compare_select_op(),
+            (int)v->compare_select_op(),
             hashOf(v->rhs()),
             hashOf(v->ret_val1()),
             hashOf(v->ret_val2())));
